Adds jest config and moves page components to their own folder
