+++ jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  testPathIgnorePatterns: ['<rootDir>/config'],
+  moduleDirectories: ['node_modules', './'],
+  moduleFileExtensions: ['js', 'jsx', 'json', 'node'],
+  testRegex: '(/__tests__/.*|(\\.|/)(test))\\.(jsx?)$',
+}
