+++ package.json
@@ -11,6 +11,7 @@
     "@rails/webpacker": "3.5",
     "axios": "^0.19.0",
     "babel-preset-react": "^6.24.1",
+    "core-js": "3",
     "dotenv": "^6.0.0",
     "immutable": "^3.8.2",
     "lodash": "^4.17.13",
@@ -28,6 +29,10 @@
     "webpack-cli": "^3.1.0"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.7.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.7.7",
+    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
+    "@babel/plugin-transform-modules-commonjs": "^7.7.5",
     "@babel/preset-env": "^7.7.7",
     "@babel/preset-react": "^7.7.4",
     "@testing-library/jest-dom": "^4.2.4",
